Add files via upload
This initializes the delta x vector for use in A*deltax = 0 with the correct nonbasic variables set to 1 or 0 when given a starting vector. It's not coded to work in Sonia's program yet, and doesn't get the final answers for the delta x vectors.  Also it assumes all 0 values in the bfs, x(1), x(2), etc... vectors are nonbasic variables.